Create event records in the background
It could be a big performance hit to always create an event before each request. One solution is to create the event record in the background.

A future optimization would be to allow a developer to chose what events should be created in the background, and what can be created inline.

Issues
------
- Closes #1